+++ db/migrate/20250403175305_create_posts.rb
@@ -1,0 +1,10 @@
+class CreatePosts < ActiveRecord::Migration[8.0]
+  def change
+    create_table :posts do |t|
+      t.string :title
+      t.text :content
+
+      t.timestamps
+    end
+  end
+end
